[Refactoring] CommandFactory.h 파일 생성 및 class 선언 이동
