+++ EmployeeManagementSystem/EmployeeManagementSystem/EmployeeManagementSystem.vcxproj
@@ -150,6 +150,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Command.h" />
+    <ClInclude Include="CommandFactory.h" />
     <ClInclude Include="Common.h" />
     <ClInclude Include="Database.h" />
     <ClInclude Include="Employee.h" />
